set stdout to None and set console output to False when start mindinsight in subprocess
If the env or gunicorn has some error, there will not be any error in console
or error log file when we set the subprocss stdout to devnull.
So I set it to None it will inhert its parent process stdout.
When gunicorn cause error, it will show error in console.