print error level message one time when failed
Change the error level of the error message, fill the detail message
when parsing failed, like crc failed.

Save the lineage_parsing when parsing failed. Because it can save the
cached status, it will skip the failed files when reloading.

As a result, it will keep the failed lineage_parser. When the diretory
is deleted, it will be deleted. But it can help print the failed message
only one time.