Added Vulkan Vision, a framework for fine-grained analysis of Vulkan pipeline execution on the GPU published at CGO 2021. Analysis includes AI workloads through the compute pipeline.
Vulkan Vision inserts dynamic instrumentation to gain unprecedented execution trace information of vulkan applications. For example, the execution trace can be used to determine the hotness of various regions of a compute shader. In addition, the runtime SIMD utilization can be cross-referenced with the hotness to find optimization candidates. Vulkan Vision was accepted to the Code Generation and Optimization 2021 conference